fix: moved validation functions into the same folder under /utils
